# docs/04-API명세-openapi.yaml
openapi: 3.0.3
info:
  title: HOOPLY API
  version: 1.0.0
  description: 농구 동호회 팀 매칭 서비스. 인증은 Bearer JWT(access 30분/refresh 14일). 모든 4xx/5xx 본문은 ErrorResponse {code, message, details[]}.
servers:
- url: https://hooply-backend.onrender.com/api/v1
  description: 운영
- url: http://localhost:8000/api/v1
  description: 로컬
security:
- bearerAuth: []
paths:
...
  /teams:
    post:
      summary: 팀 생성
      tags:
      - 팀 · 참가자
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamCreated'
...
  /teams/{team_id}:leave:
    post:
      summary: 팀 나가기
      tags:
      - 팀 · 참가자
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          description: 성공
        '422':
          description: CANNOT_DEMOTE_LAST_MANAGER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /assignments/candidates/{candidate_id}/ai-explanation:
    post:
      summary: AI 배정 설명 (매니저용)
      tags:
      - AI 설명
      parameters:
      - name: candidate_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiExplanation'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{event_id}/assignment/adopted/ai-message:
    get:
      summary: AI 한마디 (팀원용)
      tags:
      - AI 설명
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiMessage'
        '403':
          description: NOT_A_MEMBER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: NOT_ADOPTED_YET
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{event_id}/tactics/ai-recommend:
    post:
      summary: AI 전술 추천 설명
      tags:
      - AI 설명
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      - name: squad_no
        in: query
        required: true
        schema:
          type: integer
      - name: zone
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiTactics'
        '403':
          description: NOT_A_MEMBER / NOT_ATTENDEE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: NOT_ADOPTED_YET / NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users:
    get:
      summary: 전체 사용자 검색
      tags:
      - 관리자
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
          nullable: true
      - name: page
        in: query
        required: false
        schema:
          type: integer
      - name: size
        in: query
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AdminUserRow_'
  /admin/players/{player_id}/raw:
    get:
      summary: 선수 원시 데이터 열람
      tags:
      - 관리자
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerRawData'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/players/{player_id}/rating:
    patch:
      summary: 실력 지표 수동 보정
      tags:
      - 관리자
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingAdjust'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerCardDetailed'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/audit-logs:
    get:
      summary: 감사 로그 조회
      tags:
      - 관리자
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
      - name: size
        in: query
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AuditLogView_'
  /admin/teams:
    get:
      summary: 팀 목록 (승인 대기 우선)
      tags:
      - 관리자
      parameters:
      - name: approval
        in: query
        required: false
        schema:
          type: string
          nullable: true
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemList_TeamDetail_'
  /admin/teams/{team_id}:approve:
    post:
      summary: 팀 승인
      tags:
      - 관리자
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetail'
  /admin/teams/{team_id}:reject:
    post:
      summary: 팀 승인 거절
      tags:
      - 관리자
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminUserRow:
      properties:
        id:
          type: integer
        email:
          type: string
          nullable: true
        name:
          type: string
        nickname:
          type: string
          nullable: true
        global_role:
          type: string
        onboarding_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
      type: object
      required:
      - id
      - email
      - name
      - nickname
      - global_role
      - onboarding_completed
      - created_at
      - deleted_at
...
    AiExplanation:
      properties:
        summary:
          type: string
        key_players:
          items:
            type: string
          type: array
        chemistry:
          items:
            type: string
          type: array
        gaps:
          items:
            type: string
          type: array
        watch_point:
          type: string
        fallback:
          type: boolean
        text:
          type: string
          nullable: true
        cached:
          type: boolean
        fail_reason:
          type: string
          nullable: true
      type: object
      required:
      - summary
      - key_players
      - chemistry
      - gaps
      - watch_point
      - fallback
    AiMessage:
      properties:
        in_assignment:
          type: boolean
        why_position:
          type: string
        role:
          type: string
        partner:
          type: string
        fallback:
          type: boolean
        text:
          type: string
          nullable: true
        cached:
          type: boolean
        fail_reason:
          type: string
          nullable: true
      type: object
      required:
      - fallback
    AiTacticItem:
      properties:
        play_key:
          type: string
        reason:
          type: string
        key_roles:
          items:
            type: string
          type: array
        caution:
          type: string
      type: object
      required:
      - play_key
      - reason
      - key_roles
      - caution
    AiTactics:
      properties:
        squad_no:
          type: integer
        one_liner:
          type: string
        items:
          items:
            $ref: '#/components/schemas/AiTacticItem'
          type: array
        fallback:
          type: boolean
        cached:
          type: boolean
        fail_reason:
          type: string
          nullable: true
      type: object
      required:
      - squad_no
      - one_liner
      - items
      - fallback
...
    AuditLogView:
      properties:
        id:
          type: integer
        actor_user_id:
          type: integer
          nullable: true
        action:
          type: string
        target_type:
          type: string
        target_id:
          type: integer
          nullable: true
        before:
          additionalProperties: true
          type: object
          nullable: true
        after:
          additionalProperties: true
          type: object
          nullable: true
        reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      type: object
      required:
      - id
      - actor_user_id
      - action
      - target_type
      - target_id
      - before
      - after
      - reason
      - created_at
...
    ErrorDetail:
      properties:
        field:
          type: string
          nullable: true
        reason:
          type: string
        context:
          additionalProperties: true
          type: object
          nullable: true
      type: object
      required:
      - reason
    ErrorResponse:
      properties:
        code:
          type: string
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
      type: object
      required:
      - code
      - message
...
    ItemList_TeamDetail_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TeamDetail'
          type: array
      type: object
      required:
      - items
...
    PageMeta:
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        has_next:
          type: boolean
      type: object
      required:
      - page
      - size
      - total
      - has_next
    Page_AdminUserRow_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AdminUserRow'
          type: array
        meta:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - meta
    Page_AuditLogView_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AuditLogView'
          type: array
        meta:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - meta
...
    PlayerCardDetailed:
      properties:
        id:
          type: integer
        user_id:
          type: integer
          nullable: true
        kind:
          type: string
        display_name:
          type: string
        role:
          type: string
        profile_image_url:
          type: string
          nullable: true
        height_cm:
          type: integer
          nullable: true
        skill_grade:
          $ref: '#/components/schemas/SkillGrade'
          nullable: true
        primary_position:
          $ref: '#/components/schemas/Position'
          nullable: true
        playable_positions:
          items:
            $ref: '#/components/schemas/Position'
          type: array
        attendance_rate:
          type: number
          nullable: true
        skill_confidence:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
        skill_overall:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
        prior_overall:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
        skill_axes:
          additionalProperties:
            type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
            nullable: true
          type: object
        avg_margin:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
        quarters_played:
          type: integer
        attended_events:
          type: integer
      type: object
      required:
      - id
      - kind
      - display_name
      - role
...
    PlayerRawData:
      properties:
        player:
          additionalProperties: true
          type: object
        profile:
          additionalProperties: true
          type: object
          nullable: true
        survey_answers:
          items:
            additionalProperties: true
            type: object
          type: array
        lineups:
          items:
            additionalProperties: true
            type: object
          type: array
        votes_received:
          items:
            additionalProperties: true
            type: object
          type: array
        rating_history:
          items:
            additionalProperties: true
            type: object
          type: array
      type: object
      required:
      - player
      - profile
      - survey_answers
      - lineups
      - votes_received
      - rating_history
...
    Position:
      type: string
      enum:
      - PG
      - SG
      - SF
      - PF
      - C
...
    RatingAdjust:
      properties:
        skill_overall:
          anyOf:
          - type: number
            maximum: 50.0
            minimum: -50.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
        reason:
          type: string
          maxLength: 500
          minLength: 1
      type: object
      required:
      - skill_overall
      - reason
...
    SkillGrade:
      type: string
      enum:
      - A
      - B
      - C
      - D
      - E
...
    TeamCreate:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
        home_court:
          type: string
          maxLength: 100
          nullable: true
      type: object
      required:
      - name
    TeamCreated:
      properties:
        id:
          type: integer
        team_code:
          type: string
      type: object
      required:
      - id
      - team_code
    TeamDetail:
      properties:
        id:
          type: integer
        name:
          type: string
        team_code:
          type: string
        status:
          type: string
        approval_status:
          type: string
        min_members:
          type: integer
        home_court:
          type: string
          nullable: true
        owner:
          $ref: '#/components/schemas/UserSummary'
        member_count:
          type: integer
        my_role:
          $ref: '#/components/schemas/TeamRole'
          nullable: true
        my_player_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
      type: object
      required:
      - id
      - name
      - description
      - team_code
      - status
      - min_members
      - home_court
      - owner
      - member_count
      - created_at
...
    TeamRole:
      type: string
      enum:
      - MANAGER
      - PLAYER
...
    UserSummary:
      properties:
        id:
          type: integer
        name:
          type: string
        nickname:
          type: string
          nullable: true
        profile_image_url:
          type: string
          nullable: true
      type: object
      required:
      - id
      - name
...